Слияние кода из svn mpc-hc rev.5493; Обновление: ffmpeg;
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@688 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/Subtitles/STS.cpp
@@ -2866,10 +2866,10 @@ bool CSimpleTextSubtitle::SaveAs(CString fn, exttype et, double fps, CTextFile::
 		CStringW str2;
 
 		if (et == EXTSRT) {
-			str2.Format(fmt, i-k+1, hh1, mm1, ss1, ms1, hh2, mm2, ss2, ms2, str);
+			str2.Format(fmt, i - k + 1, hh1, mm1, ss1, ms1, hh2, mm2, ss2, ms2, str);
 		} else if (et == EXTSUB) {
 			str.Replace('\n', '|');
-			str2.Format(fmt, t1*fps/1000, t2*fps/1000, str);
+			str2.Format(fmt, int(t1 * fps / 1000), int(t2 * fps / 1000), str);
 		} else if (et == EXTSMI) {
 			str.Replace(L"\n", L"<br>");
 			str2.Format(fmt, t1, str, t2);
@@ -2879,8 +2879,8 @@ bool CSimpleTextSubtitle::SaveAs(CString fn, exttype et, double fps, CTextFile::
 		} else if (et == EXTSSA) {
 			str.Replace(L"\n", L"\\N");
 			str2.Format(fmt,
-						hh1, mm1, ss1, ms1/10,
-						hh2, mm2, ss2, ms2/10,
+						hh1, mm1, ss1, ms1 / 10,
+						hh2, mm2, ss2, ms2 / 10,
 						TToW(stse.style), TToW(stse.actor),
 						stse.marginRect.left, stse.marginRect.right, (stse.marginRect.top + stse.marginRect.bottom) / 2,
 						TToW(stse.effect), str);
@@ -2888,8 +2888,8 @@ bool CSimpleTextSubtitle::SaveAs(CString fn, exttype et, double fps, CTextFile::
 			str.Replace(L"\n", L"\\N");
 			str2.Format(fmt,
 						stse.layer,
-						hh1, mm1, ss1, ms1/10,
-						hh2, mm2, ss2, ms2/10,
+						hh1, mm1, ss1, ms1 / 10,
+						hh2, mm2, ss2, ms2 / 10,
 						TToW(stse.style), TToW(stse.actor),
 						stse.marginRect.left, stse.marginRect.right, (stse.marginRect.top + stse.marginRect.bottom) / 2,
 						TToW(stse.effect), str);
